backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]
requests
sqlmodel
psycopg[binary]
langchain
langgraph
langchain-openai
langchain-groq
langchain-google-genai
langgraph-supervisor
openai
sqlalchemy[asyncio]==2.0.35
sse-starlette==2.0.0
google-generativeai>=0.5.3


tenacity==8.2.3
PyYAML==6.0.2